Moved base64 implementation to another file
