+++ src/base64.hpp
@@ -16,9 +16,10 @@
 
 #pragma once
 
-#include "bot.hpp"
+#include "string"
 
-namespace aelliptic { namespace commands {
-    void base64(TgBot::Message::Ptr message);
-}}
-
+namespace aelliptic {
+    std::string base64_decode(std::string const& encoded_string);
+    std::string base64_encode(unsigned char const* bytes_to_encode, 
+                              unsigned int in_len);
+}
